Improve conditionals to reduce failures
The conditionals based on nspawn_config_exists.stat.exists
fail in some conditions. When the task doing the state does
not execute, there is no dictionary result with the appropriate
structure, causing tasks relying on the structure to fail.

This patch re-arranges the conditionals so that when they
are evaluated, they do so in a better process of elimination.
First, evaluate the same conditional as the stat task, then
validate that the stat dict exists, then check the key in
the stat result.

Change-Id: I6f2d8943c4279edf44907aa145935be858dbc673