Use networkd role to setup test networks
The test networks were being setup using networkd and a custom template,
this change updates the process so we're dog fooding.

Additionally change the container config drop so that its executing a
container restart when needed and in the right order.

Several tests were removed because they are basic operations which are
already being tested in the nspawn_hosts role.

Change-Id: I50799d53f8d75f043ed12d4e50da3d55a4ee159a
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>