Correct test register
This change corrects the name of a register used to cleanup containers
when running tests multiple times. While this register would not harm a
normal run, it does cause issues when re-running the tests due to the
fact the variable is used to determine if a container is removed and it
is undefined at this point.

Change-Id: I4e0c07c4047ccdde3becfb7345d2564bef0489da
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>