Update qgroup setup to be more explicit
The default qgroup setup was setting the qgroup limits to none which
allows the system to dynamical expand or contract. While this works
there can be a lag between the time it takes to grow vs the time it
takes to hit the limit. This change sets the limits accordingly should a
deployer not have a container specific limit set already.

Change-Id: Iea2498bfe1c391f4b4bee295c8f04b2655f404d9