Add test for host_only networks and ensure network ordering
The `host_only` network option did not have a functional test, this
change adds one so that the mv-extra interface will now run with this
which will ensure its functional when enabled. The "nspawn_combined_networks"
also had an issue where the first network in the container config could
be something other than the nspawn container network. This would shift
the gateway from the mv-nspawn and to whatever was defined first. A
change to the "nspawn_combined_networks" will now ensure
"nspawn_networks" is referenced first.

Depends-On: Ic280db301ce44b98ac35c8911999dec5fa263a29
Change-Id: Id20afab8b705f6528ba5575ee4704c1661e6e65c
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>